requirements.txt
slack-bolt
python-dotenv
fastapi
uvicorn
google-generativeai
